% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countiesinstates.R
\name{countiesinstates}
\alias{countiesinstates}
\title{countiesinstates
Returns all the counties found in the specified states}
\usage{
countiesinstates(states, allcounties)
}
\arguments{
\item{states}{Full capitalized state names like Maryland}

\item{allcounties}{from getcountyfullnames()}
}
\description{
countiesinstates
Returns all the counties found in the specified states
}
