% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidDownload.R
\name{covidDownload}
\alias{covidDownload}
\title{covidDownload}
\usage{
covidDownload(
  fileurl = "https://github.com/nytimes/covid-19-data/raw/master/us-counties.csv",
  writecsv = FALSE,
  localpath = getwd()
)
}
\arguments{
\item{localpath}{}
}
\value{

}
\description{
Read daily update of US county-resolution COVID data from NYT dataset on github
}
