% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotContagious.R
\name{covidPlotContagious}
\alias{covidPlotContagious}
\title{covidPlotContagious}
\usage{
covidPlotContagious(
  x,
  countylist = NULL,
  ndays,
  dayscontagious = 14,
  perx = 1e+05,
  add = FALSE,
  show = TRUE,
  ylim = NULL,
  showlinesforpriordate = TRUE,
  ...
)
}
\arguments{
\item{x}{from covidDownload}

\item{countylist}{vector of county names like "Montgomery County, Maryland" as found in unique fullname entries in covidDownload()}

\item{ndays}{show only the last ndays days of data}

\item{dayscontagious}{how many recent days of new cases to add together, as a way to approximate how many people are currently still contagious}

\item{perx}{per how many, such as per 100 or per 100000}

\item{add}{adding to existing plot or no}

\item{show}{whether to draw plot or just return info}

\item{ylim}{as in plot()}

\item{showlinesforpriordate}{draw short line fit slopes to show midpoint of linefit to each of several recent sets of days}

\item{...}{passed to plot()}
}
\description{
covidPlotContagious
}
