% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotContagiousEach.R
\name{covidPlotContagiousEach}
\alias{covidPlotContagiousEach}
\title{covidPlotContagiousEach}
\usage{
covidPlotContagiousEach(
  x,
  countylist = c("Montgomery County, Maryland",
    "District of Columbia, District of Columbia"),
  ndays,
  dayscontagious = 14,
  perx = 1e+05,
  digits = 3,
  crop_length_legend = 10,
  ...
)
}
\arguments{
\item{x}{from covidDownload}

\item{countylist}{vector of county names like "Montgomery County, Maryland" as found in unique(covidDownload()\link{ , "fullname"})}

\item{ndays}{show only the last ndays days of data}

\item{dayscontagious}{how many recent days of new cases to add together, as a way to approximate how many people are currently still contagious}

\item{perx}{per how many people such as per 100 or per 100000}

\item{digits}{controls number of significant digits in legend}

\item{crop_length_legend}{how many characters to use of the place names in legend}

\item{...}{passed to covidPlotContagious()}
}
\description{
covidPlotContagiousEach
}
