% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotCum_v_New.R
\name{covidPlotCum_v_New}
\alias{covidPlotCum_v_New}
\title{covidPlotCum_v_New}
\usage{
covidPlotCum_v_New(x, countylist = NULL, averagingtime = 3, ndays, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
covidPlotCum_v_New
}
