% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotTrend.R
\name{covidPlotTrend}
\alias{covidPlotTrend}
\title{covidPlotTrend}
\usage{
covidPlotTrend(x, countylist = NULL, ndays, ...)
}
\arguments{
\item{x}{from covidDownload()}

\item{ndays}{Filter to only the last n days of data}

\item{...}{other parameters to pass to plot()}
}
\description{
Draws plot of cumulative total cases (per 100 population) vs date
for specified US counties
}
