% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotTrendNew.R
\name{covidPlotTrendNew}
\alias{covidPlotTrendNew}
\title{covidPlotTrendNew}
\usage{
covidPlotTrendNew(
  x,
  countylist = NULL,
  averagingtime = 3,
  smoothspan,
  ndays,
  lastn = 14,
  fitlwd = 3,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
covidPlotTrendNew
}
