% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotTrendNewEachPlace.R
\name{covidPlotTrendNewEachPlace}
\alias{covidPlotTrendNewEachPlace}
\title{covidPlotTrendNewEachPlace}
\usage{
covidPlotTrendNewEachPlace(
  x,
  countylist = c("Montgomery County, Maryland",
    "District of Columbia, District of Columbia"),
  averagingtime = 3,
  smoothspan,
  ndays,
  lastn = 14,
  percap = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
covidPlotTrendNewEachPlace
}
