% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotTrendNewRunFit.R
\name{covidPlotTrendNewRunFit}
\alias{covidPlotTrendNewRunFit}
\title{covidPlotTrendNewRunFit}
\usage{
covidPlotTrendNewRunFit(
  x,
  countylist = NULL,
  averagingtime = 3,
  ndays,
  lastn = 14,
  fitlwd = 3,
  fitlty = 1,
  autoregress = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
covidPlotTrendNewRunFit
}
