% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotWorstFew.R
\name{covidPlotWorstFew}
\alias{covidPlotWorstFew}
\title{covidPlotWorstFew}
\usage{
covidPlotWorstFew(xnow, n = 7, asofhere, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
covidPlotWorstFew
}
