% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidTableSelected.R
\name{covidTableSelected}
\alias{covidTableSelected}
\title{covidTableSelected}
\usage{
covidTableSelected(xnow, asofhere, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
covidTableSelected
}
