% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidToday.R
\name{covidToday}
\alias{covidToday}
\title{covidToday}
\usage{
covidToday(x, mydate = NULL, writecsv = FALSE, localpath = getwd())
}
\arguments{
\item{x}{from covidDownload}

\item{mydate}{what date to use, defaults to most recent one in dataset}

\item{writecsv}{logical, optional, whether to write file to disk}

\item{localpath}{optional path where to write file}
}
\description{
covidToday
}
