% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtrackingDownload.R
\name{covidtrackingDownload}
\alias{covidtrackingDownload}
\title{covidtrackingDownload
Get daily data from \code{\url{'https://covidtracking.com/api/v1/states/daily.csv'}}}
\usage{
covidtrackingDownload()
}
\value{

}
\description{
covidtrackingDownload
Get daily data from \code{\url{'https://covidtracking.com/api/v1/states/daily.csv'}}
}
