% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidWrite.R
\name{covidWrite}
\alias{covidWrite}
\title{covidWrite}
\usage{
covidWrite(x, localpath = getwd())
}
\arguments{
\item{x}{from covidDownload}

\item{localpath}{local path to hard drive to store written file}
}
\description{
covidWrite
}
