% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcountyfullnames.R
\name{getcountyfullnames}
\alias{getcountyfullnames}
\title{getcountyfullnames}
\usage{
getcountyfullnames(fips)
}
\arguments{
\item{fips}{Census FIPS code as 5 digits for county}
}
\value{

}
\description{
getcountyfullnames
}
