% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latestcontagiouspercap.R
\name{latestcontagiouspercap}
\alias{latestcontagiouspercap}
\title{latestcontagiouspercap}
\usage{
latestcontagiouspercap(x, countylist = NULL, dayscontagious = 14)
}
\arguments{
\item{x}{from covidDownload}

\item{countylist}{vector of county names like "Montgomery County, Maryland" as found in unique(covidDownload()\link{ , "fullname"})}

\item{dayscontagious}{how many recent days of new cases to add together, as a way to approximate how many people are currently still contagious}
}
\description{
latestcontagiouspercap
}
