% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead.zeroes_covid.R
\name{lead.zeroes_covid}
\alias{lead.zeroes_covid}
\title{lead.zeroes_covid}
\usage{
lead.zeroes_covid(fips, length.desired)
}
\arguments{
\item{fips}{Census FIPS such as for state-county}

\item{length.desired}{how many character after leading zeroes added}
}
\value{

}
\description{
lead.zeroes_covid
}
