% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_from_cum_by_fip.R
\name{new_from_cum_by_fip}
\alias{new_from_cum_by_fip}
\title{For each subset (place), use cumulative case data to calculate new cases per day}
\usage{
new_from_cum_by_fip(cum, fip, date)
}
\arguments{
\item{cum}{Cumulative cases as of date}

\item{fip}{place, such as FIPS code from Census}

\item{date}{Dates that can be sorted, ideally with no gaps in dates for a given place}
}
\description{
Problem if more than one place has NA as FIPS code. covidDownload handled that before calling this.
Assumes dates are consecutive (no breaks) but need not be sorted by date beforehand.
If there is a gap, the new cases is the change from the prior date, whether that was
one day prior or longer. In other words, it is new cases since the last available date
for that location.
}
