% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderbylatest.R
\name{orderbylatest}
\alias{orderbylatest}
\title{orderbylatest returns ranks of places based on data on most recent day only}
\usage{
orderbylatest(myvar, placevar, mydate)
}
\arguments{
\item{myvar}{vector of data}

\item{placevar}{vector of county names}

\item{mydate}{vector of dates where you want just most recent}
}
\description{
1 = lowest value is rank of 1
returns number that is new sort order for
dataset with multiple places and multiple dates,
based on rank of place-specific myvar on latest date
}
