% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stillcontagious.R
\name{stillcontagious}
\alias{stillcontagious}
\title{For one place, calculate cumulative total new cases for last several days}
\usage{
stillcontagious(new, dayscontagious)
}
\arguments{
\item{new}{vector of new cases each day}

\item{dayscontagious}{how many recent days to add up}
}
\description{
Assumes data are sorted in date order oldest to newest
see related code in covidPlotContagious
}
