% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stillcontagious_percap_bycounty.R
\name{stillcontagious_percap_bycounty}
\alias{stillcontagious_percap_bycounty}
\title{Calculate cumulative recent new cases per capita for each place}
\usage{
stillcontagious_percap_bycounty(
  date,
  new,
  countyname,
  pop,
  dayscontagious = 14
)
}
\arguments{
\item{date}{dates (not necessarily sorted)}

\item{new}{new cases per day for given place and date}

\item{countyname}{place names (defining subsets within each of which the running cum recent totals are gotten)}

\item{pop}{population count for given place}

\item{dayscontagious}{how many recent days of new cases to add up}
}
\description{
Calculate cumulative recent new cases per capita for each place
}
\examples{
na
}
