% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_counties.R
\name{sum_counties}
\alias{sum_counties}
\title{sum_counties}
\usage{
sum_counties(
  df,
  countylist = NULL,
  countycolname = "fullname",
  datecolname = "date"
)
}
\arguments{
\item{datecolname}{}
}
\value{

}
\description{
sum_counties
}
