% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidDownload.R
\name{covidDownload}
\alias{covidDownload}
\title{covidDownload}
\usage{
covidDownload(
  fileurl = "https://github.com/nytimes/covid-19-data/raw/master/us-counties.csv",
  writecsv = FALSE,
  localpath = getwd(),
  testing = FALSE
)
}
\arguments{
\item{fileurl}{optional but will not really work if csv is not formatted as the one here}

\item{writecsv}{optional logical}

\item{localpath}{optional}

\item{testing}{whether to print extra info to console}
}
\description{
Read daily update of US county-resolution COVID data from NYT dataset on github
}
\examples{
  x <- covidDownload()
  covidPlotBarNow(x)
}
