% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotBarNow.R
\name{covidPlotBarNow}
\alias{covidPlotBarNow}
\title{covidPlotBarNow}
\usage{
covidPlotBarNow(
  x,
  asofhere,
  horiz = FALSE,
  n = 10,
  noworever = "ever",
  dayscontagious = 14,
  inapp = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
covidPlotBarNow
}
