% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotContagious.R
\name{covidPlotContagious}
\alias{covidPlotContagious}
\title{covidPlotContagious}
\usage{
covidPlotContagious(
  x,
  countylist = NULL,
  ndays,
  dayscontagious = 14,
  add = FALSE,
  show = TRUE,
  ylim = NULL,
  showlinesforpriordate = TRUE,
  ...
)
}
\arguments{
\item{x}{from covidDownload}

\item{countylist}{vector of county names like "Montgomery County, Maryland" as found in unique(covidDownload()\link{ , "fullname"})}

\item{ndays}{show only the last ndays days of data}

\item{dayscontagious}{how many recent days of new cases to add together, as a way to approximate how many people are currently still contagious}

\item{...}{passed to plot()}
}
\description{
covidPlotContagious
}
