% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotContagious.R
\name{covidPlotContagious}
\alias{covidPlotContagious}
\title{covidPlotContagious}
\usage{
covidPlotContagious(
  x,
  countylist = NULL,
  ndays,
  dayscontagious = 14,
  add = FALSE,
  show = TRUE,
  ylim = NULL,
  showlinesforpriordate = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
covidPlotContagious
}
