% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotContagiousEach.R
\name{covidPlotContagiousEach}
\alias{covidPlotContagiousEach}
\title{covidPlotContagiousEach}
\usage{
covidPlotContagiousEach(
  x,
  countylist = c("Montgomery County, Maryland",
    "District of Columbia, District of Columbia"),
  ndays,
  dayscontagious = 14,
  digits = 3,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
covidPlotContagiousEach
}
