% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotTrend.R
\name{covidPlotTrend}
\alias{covidPlotTrend}
\title{covidPlotTrend}
\usage{
covidPlotTrend(x, countylist = NULL, perx = 1e+05, ndays, ...)
}
\arguments{
\item{x}{from covidDownload()}

\item{countylist}{vector of county names like "Montgomery County, Maryland" as found in unique(covidDownload()\link{ , "fullname"})}

\item{ndays}{show only the last ndays days of data}

\item{...}{other parameters to pass to plot()}
}
\description{
Draws plot of cumulative total cases (per perx population) vs date
for specified US counties
}
