% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlotWorstFew.R
\name{covidPlotWorstFew}
\alias{covidPlotWorstFew}
\title{covidPlotWorstFew}
\usage{
covidPlotWorstFew(xnow, n = 7, asofhere, ...)
}
\arguments{
\item{xnow}{from covidToday(covidDownload())}

\item{n}{how many places, optional}

\item{asofhere}{date to use, optional}

\item{...}{passed to barplot}
}
\description{
covidPlotWorstFew
}
