% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidTableWorstFew.R
\name{covidTableWorstFew}
\alias{covidTableWorstFew}
\title{covidTableWorstFew}
\usage{
covidTableWorstFew(xnow, n = 7, asofhere)
}
\arguments{
\item{xnow}{from covidToday(covidDownload())}

\item{n}{how many places, optional}

\item{asofhere}{what date, optional}
}
\description{
covidTableWorstFew
}
