% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latestcontagiouspercap.R
\name{latestcontagiouspercap}
\alias{latestcontagiouspercap}
\title{latestcontagiouspercap}
\usage{
latestcontagiouspercap(x, countylist = NULL, dayscontagious = 14)
}
\arguments{
\item{dayscontagious}{}
}
\value{

}
\description{
latestcontagiouspercap
}
