% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stillcontagious.R
\name{stillcontagious}
\alias{stillcontagious}
\title{stillcontagious}
\usage{
stillcontagious(new, dayscontagious)
}
\arguments{
\item{dayscontagious}{how many recent days of new cases to add together, as a way to approximate how many people are currently still contagious}
}
\description{
stillcontagious
}
