% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_counties.R
\name{sum_counties}
\alias{sum_counties}
\title{sum_counties}
\usage{
sum_counties(
  df,
  countylist = NULL,
  countycolname = "fullname",
  datecolname = "date"
)
}
\arguments{
\item{df}{data.frame from covidDownload()}

\item{countylist}{vector of county names like "Montgomery County, Maryland" as found in unique fullname entries in covidDownload()}

\item{countycolname}{optional name of data.frame column with county names}

\item{datecolname}{optional name of data.frame column with date}
}
\description{
Create appropriately aggregated statistics on cases etc. from all counties in the data.frame
}
